requirements.txt
Django==3.2
# django-widget-tweaks==1.4.8
djangorestframework==3.12.4
django-extensions>=3.1.3
ipython>=7.22.0
ipdb
# pytz==2017.3
